portfolio-webapp/backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
jupyterlab==4.3.5
PyJWT==2.8.0
pyotp==2.9.0
bcrypt==4.1.3
qrcode[pil]==7.4.2
